+++ mymocore/Core/Traits/PostTypeHookAction.php
@@ -67,6 +67,12 @@ trait PostTypeHookAction
                     'position' => $iargs->get('menu_position')
                 ]
             );
+
+            $this->registerPermalink($taxonomy, [
+                'label' => $iargs->get('label'),
+                'base' => $iargs->get('singular'),
+                'callback' => 'Mymo\\Theme\\Http\\Controllers\\TaxonomyController'
+            ]);
         }
     }
