--- mymocore/Core/resources/views/components/form_taxonomy.blade.php
@@ -1,1 +1,0 @@
-@livewire('tadcms::component.form-taxonomy', ['taxonomy' => 'category'])
