+++ mymocore/Theme/routes/web.php
@@ -12,8 +12,6 @@
  * Time: 3:22 PM
  */
 
-Route::get('/', 'HomeController@index')->name('home');
-
 Route::get('/search', 'SearchController@index')->name('search');
 
 Route::get('/login', 'HomeController@index')->name('login');
@@ -22,4 +20,9 @@ Route::get('/register', 'HomeController@index')->name('register');
 
 Route::get('/{slug}', 'PageController@index')
     ->name('frontend')
+    ->where('slug', '[a-z0-9\-]+');
+
+Route::get('/{base}/{slug}', 'RouteController@index')
+    ->name('frontend')
+    ->where('slug', '[a-z0-9\-\/]+')
     ->where('slug', '[a-z0-9\-\/]+');
